+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes
 * 2021-06-14 - **Allow scoring completion of features/activities in block settings**
 * 2021-06-14 - Changes in the layout of the ranking list
 * 2021-06-14 - Update of the game help
+* 2021-07-20 - **Possibility of download the ranking list**
 
 
 ### Changes Version 1.20 Version for Moodle 3.5 onwards